use build.rs instead of macro
